# PERF: vectorize _interp_limit