--- a/pandas/core/missing.py
+++ b/pandas/core/missing.py
@@ -634,13 +634,6 @@ def fill_zeros(result, x, y, name, fill):
     return result
 
 
-def _interp_limit0(invalid, fw_limit, bw_limit):
-    "Get idx of values that won't be filled b/c they exceed the limits."
-    for x in np.where(invalid)[0]:
-        if invalid[max(0, x - fw_limit):x + bw_limit + 1].all():
-            yield x
-
-
 def _interp_limit(invalid, fw_limit, bw_limit):
     """Get idx of values that won't be filled b/c they exceed the limits.