xref https://github.com/pandas-dev/pandas/issues/16584

Here are the timings vs. master

```
       before           after         ratio
     [ff0d1f4d]       [6b1552f1]
-      12.3±0.3ms      2.28±0.04ms     0.18  frame_methods.Interpolate.time_interpolate_some_good_infer
-      12.6±0.7ms      1.15±0.01ms     0.09  frame_methods.Interpolate.time_interpolate_some_good
-           1.30s         53.8±1ms     0.04  frame_methods.Interpolate.time_interpolate

```

I still need to time vs 0.20.1 (will post those timings later, have to run for a bit now). I may have one more spot to optimize.