@TomAugspurger 2017-06-03
Also, the master and 0.20.1 versions scale poorly with `limit`, but that's not benchmarked anywhere. Here's some timings

limit / version | 0.20.1  | master | interp-perf
--------------- | ------- | ------ | -----------
None            | 42.2 ms | 1.37 s | 43.8 ms
50              | 1.45 s  | 1.33 s | 66.2 ms
500             | 1.40  s  | 1.36 s | 62.5 ms

So we're slightly slower than 0.20.1 for the default, but much faster with any limit (notice the units).

I'll add a benchmark for `limit` and `direction`. Anyone know if there's a way to run new benchmarks against old commits?