@TomAugspurger 2017-06-03
Oh sorry that's not used anymore. Just left it around for testing that my new implementation was the same. I'll remove it. 

> On Jun 3, 2017, at 13:17, Jeff Reback <notifications@github.com> wrote:
> 
> @jreback commented on this pull request.
> 
> In pandas/core/missing.py:
> 
> > @@ -630,3 +632,65 @@ def fill_zeros(result, x, y, name, fill):
>              result = result.reshape(shape)
>  
>      return result
> +
> +
> +def _interp_limit0(invalid, fw_limit, bw_limit):
> +    "Get idx of values that won't be filled b/c they exceed the limits."
> +    for x in np.where(invalid)[0]:
> +        if invalid[max(0, x - fw_limit):x + bw_limit + 1].all():
> should rewrite this as well
> 
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub, or mute the thread.
>