@jreback 2017-06-04
no this is all fine maybe make an issue for 0.21 to have a look